Fix websocket channel disconnection after component unmount.
`socketClient` channel isn't disconnected on component unmount. By adding a `.off()` method on `useEffect` callbacks we can make sure that the connection is closed after a component is unmounted to avoid leaving connections hanging.

Fixed as well state update for messages with a functional update method. The `useEffect` containing the socket connection was being triggered on every render (undesired behavior).